+++ src/lib/hooks/use-locations.ts
@@ -1,17 +1,18 @@
-import { useEffect, useState } from 'react';
-import { fetcher } from "src/lib/utilityFunctions";
-import type { Location } from 'src/lib/database/location';
+import { useEffect, useState } from "react";
+import { fetcher } from "~/lib/utils/fetch";
+import type { Location } from "~/lib/database/location";
 
 export const useLocations = (url: string) => {
-  const [locations, setLocations] = useState<Array<Location> | undefined>(undefined);
+  const [locations, setLocations] = useState<Array<Location> | undefined>(
+    undefined
+  );
 
   useEffect(() => {
     const fetchLocations = async () => {
       try {
         const locations: Location[] = await fetcher(url);
         setLocations(locations);
-      }
-      catch (e) {
+      } catch (e) {
         console.error(`useLocations::`, e);
         setLocations([]);
       }
