+++ src/lib/hooks/use-sensor-types.ts
@@ -1,16 +1,17 @@
-import { useEffect, useState } from 'react';
-import { fetcher } from "src/lib/utilityFunctions";
+import { useEffect, useState } from "react";
+import { fetcher } from "~/lib/utils/fetch";
 
 export const useSensorTypes = (url: string) => {
-  const [sensorTypes, setSensorTypes] = useState<Array<string> | undefined>(undefined);
+  const [sensorTypes, setSensorTypes] = useState<Array<string> | undefined>(
+    undefined
+  );
 
   useEffect(() => {
     const fetchSensorTypes = async () => {
       try {
         const types: string[] = await fetcher(url);
         setSensorTypes(types);
-      }
-      catch (e) {
+      } catch (e) {
         console.error(`useSensorTypes::`, e);
         setSensorTypes([]);
       }
